Programming Assignment 2 Coursera
In the second assignment of coursera lessons about R programming we have to create a function that caches the inverse of a matrix .

In the first function makeCacheMatrix we are setting value to a special matrix and getting the value. After that we set value to the inverse matreix and get that value.

In the second function cacheSolve if there is not any value for the inverse of the matrix we calculate it throught the solve function otherwise we return the value from the cache wich allready exists.